@@ -9,8 +9,8 @@ documentation = "https://docs.rs/reqwest"
 authors = ["Sean McArthur <sean@seanmonstar.com>"]
 readme = "README.md"
 license = "MIT OR Apache-2.0"
-edition = "2018"
-rust-version = "1.57.0"
+edition = "2021"
+rust-version = "1.63.0"
 autotests = true
 
 [package.metadata.docs.rs]
