@@ -2,6 +2,7 @@
 mod support;
 
 use futures_util::stream::StreamExt;
+use support::delay_server;
 use support::server;
 
 #[cfg(feature = "json")]
@@ -180,8 +181,7 @@ async fn overridden_dns_resolution_with_gai() {
 
     let overridden_domain = "rust-lang.org";
     let url = format!(
-        "http://{}:{}/domain_override",
-        overridden_domain,
+        "http://{overridden_domain}:{}/domain_override",
         server.addr().port()
     );
     let client = reqwest::Client::builder()
@@ -203,8 +203,7 @@ async fn overridden_dns_resolution_with_gai_multiple() {
 
     let overridden_domain = "rust-lang.org";
     let url = format!(
-        "http://{}:{}/domain_override",
-        overridden_domain,
+        "http://{overridden_domain}:{}/domain_override",
         server.addr().port()
     );
     // the server runs on IPv4 localhost, so provide both IPv4 and IPv6 and let the happy eyeballs
@@ -238,8 +237,7 @@ async fn overridden_dns_resolution_with_trust_dns() {
 
     let overridden_domain = "rust-lang.org";
     let url = format!(
-        "http://{}:{}/domain_override",
-        overridden_domain,
+        "http://{overridden_domain}:{}/domain_override",
         server.addr().port()
     );
     let client = reqwest::Client::builder()
@@ -263,8 +261,7 @@ async fn overridden_dns_resolution_with_trust_dns_multiple() {
 
     let overridden_domain = "rust-lang.org";
     let url = format!(
-        "http://{}:{}/domain_override",
-        overridden_domain,
+        "http://{overridden_domain}:{}/domain_override",
         server.addr().port()
     );
     // the server runs on IPv4 localhost, so provide both IPv4 and IPv6 and let the happy eyeballs
@@ -407,4 +404,101 @@ fn update_json_content_type_if_set_manually() {
         .expect("request is not valid");
 
     assert_eq!("application/json", req.headers().get(CONTENT_TYPE).unwrap());
+}
+
+#[cfg(all(feature = "__tls", not(feature = "rustls-tls-manual-roots")))]
+#[tokio::test]
+async fn test_tls_info() {
+    let resp = reqwest::Client::builder()
+        .tls_info(true)
+        .build()
+        .expect("client builder")
+        .get("https://google.com")
+        .send()
+        .await
+        .expect("response");
+    let tls_info = resp.extensions().get::<reqwest::tls::TlsInfo>();
+    assert!(tls_info.is_some());
+    let tls_info = tls_info.unwrap();
+    let peer_certificate = tls_info.peer_certificate();
+    assert!(peer_certificate.is_some());
+    let der = peer_certificate.unwrap();
+    assert_eq!(der[0], 0x30); // ASN.1 SEQUENCE
+
+    let resp = reqwest::Client::builder()
+        .build()
+        .expect("client builder")
+        .get("https://google.com")
+        .send()
+        .await
+        .expect("response");
+    let tls_info = resp.extensions().get::<reqwest::tls::TlsInfo>();
+    assert!(tls_info.is_none());
+}
+
+// NOTE: using the default "curernt_thread" runtime here would cause the test to
+// fail, because the only thread would block until `panic_rx` receives a
+// notification while the client needs to be driven to get the graceful shutdown
+// done.
+#[tokio::test(flavor = "multi_thread", worker_threads = 2)]
+async fn highly_concurrent_requests_to_http2_server_with_low_max_concurrent_streams() {
+    let client = reqwest::Client::builder()
+        .http2_prior_knowledge()
+        .build()
+        .unwrap();
+
+    let server = server::http_with_config(
+        move |req| async move {
+            assert_eq!(req.version(), http::Version::HTTP_2);
+            http::Response::default()
+        },
+        |builder| builder.http2_only(true).http2_max_concurrent_streams(1),
+    );
+
+    let url = format!("http://{}", server.addr());
+
+    let futs = (0..100).map(|_| {
+        let client = client.clone();
+        let url = url.clone();
+        async move {
+            let res = client.get(&url).send().await.unwrap();
+            assert_eq!(res.status(), reqwest::StatusCode::OK);
+        }
+    });
+    futures_util::future::join_all(futs).await;
+}
+
+#[tokio::test]
+async fn highly_concurrent_requests_to_slow_http2_server_with_low_max_concurrent_streams() {
+    let client = reqwest::Client::builder()
+        .http2_prior_knowledge()
+        .build()
+        .unwrap();
+
+    let server = delay_server::Server::new(
+        move |req| async move {
+            assert_eq!(req.version(), http::Version::HTTP_2);
+            http::Response::default()
+        },
+        |mut http| {
+            http.http2_only(true).http2_max_concurrent_streams(1);
+            http
+        },
+        std::time::Duration::from_secs(2),
+    )
+    .await;
+
+    let url = format!("http://{}", server.addr());
+
+    let futs = (0..100).map(|_| {
+        let client = client.clone();
+        let url = url.clone();
+        async move {
+            let res = client.get(&url).send().await.unwrap();
+            assert_eq!(res.status(), reqwest::StatusCode::OK);
+        }
+    });
+    futures_util::future::join_all(futs).await;
+
+    server.shutdown().await;
 }
